Set raffle duration and check for picking winner
Create an immutable variable to save the minimum duration of the raffle
and set it in the constructor.
Create an immutable variable to save the last timestamp and set it in
the constructor.
Set the pickWinner() function to check that the minimum duration for the
raffle has been met or revert the call.